cargo clippy -p jolt-core --fix

+++ jolt-core/src/jolt/subtable/sll.rs
@@ -61,7 +61,7 @@ impl<F: PrimeField, const CHUNK_INDEX: usize, const WORD_SIZE: usize> LassoSubta
 
         // min with 1 << b is included for test cases with subtables of bit-length smaller than 6
         for k in 0..min(WORD_SIZE, 1 << b) {
-            let k_bits = (k as usize)
+            let k_bits = k
                 .get_bits(log_WORD_SIZE)
                 .iter()
                 .map(|bit| F::from_u64(*bit as u64).unwrap())
